+++ asapdiscovery-data/asapdiscovery/data/schema_v2/ligand.py
@@ -13,65 +13,13 @@ from asapdiscovery.data.openeye import (
     smiles_to_oemol,
 )
 from asapdiscovery.data.schema import ExperimentalCompoundData
-from pydantic import BaseModel, ByteSize, Field, validator
+from pydantic import  Field
 
+
+from .schema_base import DataModelAbstractBase
 
 class InvalidLigandError(ValueError):
     ...
-
-
-class DataStorageType(str, Enum):
-    sdf = "sdf"
-    pdb = "pdb"
-
-
-def read_file_directly(file: str | Path) -> str:
-    with open(str(file)) as f:
-        contents = f.read()
-    return contents
-
-
-def write_file_directly(file: str | Path, data: str) -> None:
-    with open(str(file), "w") as f:
-        f.write(data)
-
-
-def utf8len(s: str) -> int:
-    return len(s.encode("utf-8"))
-
-
-class DataModelAbstractBase(BaseModel):
-    """
-    Base class for asapdiscovery pydantic models that simplify dictionary, JSON
-    and other behaviour
-    """
-
-    @classmethod
-    def from_dict(cls, dict):
-        return cls.parse_obj(dict)
-
-    @classmethod
-    def from_json(cls, json_str):
-        return cls.parse_obj(json.loads(json_str))
-
-    # check_fields required to be false as `data` not defined for baseclass
-    @validator("data", check_fields=False)
-    def data_must_not_be_empty(cls, v):
-        if not v:
-            raise ValueError("Data field can not be empty")
-        return v
-
-    @property
-    def size(self) -> ByteSize:
-        """Size of the resulting JSON object for this class"""
-        return ByteSize(utf8len(self.json())).human_readable()
-
-    def data_equal(self, other: DataModelAbstractBase) -> bool:
-        return self.data == other.data
-
-    class Config:
-        validate_assignment = True
-
 
 # Ligand Schema
